Blockchain: Maintain chain state and process incoming blocks
We maintain a single ChainState corresponding to the current tip. If a
new block comes in, and it's a valid chain that is longer than the
current tip, we switch to it. If a block comes in whose parent is not
yet known, we request the parent from the network (not implemented
yet).