better error message if this happens
std::time::Instant does not guarantee to track time while system is suspended
or hybernating. If the system suspend for 30min the Instant won't notice it.

In future we might want to handle such thing and there fore to compare the scheduled
event to detect this or at least to detect the system events. This can be detected on
windows C#. Might be possible to do the same on other systems.