Make bump_peer_for_block_fetch miss non-fatal
In network::p2p::comm::PeerMap::bump_peer_for_block_fetch,
don't panic if the node to bump is no longer in the map.
This can happen in the wild if an announcement is processed
after the peer comms entry is removed from the map due to
(probably) a subscription stream having closed.