Properly filter checkpoints for given descendant
When requested a range from checkpoints to an end block,
only consider those of the checkpoints that are actually the
ancestors of the end block. This is because the requester may
be the remote peer restoring the chain from checkpoints picked
among known blocks in its storage and a newly announced block, which
may be a descendant of only some of those due to unresolved
forks in the chain.