Remove the cap on number of network blocks
The limit is semi-arbitrary, and for preventing potential DoS from
a rogue node, the blockchain validation protocol should be enough
in stopping any rogue nodes. For out-of-order blocks that go to
the candidate forest, an entry from a previously validated
header chain pull must exist, so when we implement proper validation
for header chains, that hole will be closed too.