fix rlimit builds on different libc impls
rlimit may be using different return or argument types depending on a
particular libc implementation. Due to that, getrlimit now accepts the
RlimitResource enum, which is internally converted to a correct type
inferred from the libc implementation being used.

In addition to that, the return type of libc::getrlimit is always
converted to u64.