Embed CI test hooks permanently
Reserve special branch name ci/test/nightly to test the nightly workflow
by pushing to the branch.
Relax the repository owner checks by allowing runs in any repository on
pushing a commit that starts with 'TEST CI:'.
The release workflow can be triggered in forks by making such a commit
and tagging it with a 'v*' tag.