blockchain: check rewards distribution before applying a block
Block rewards distribution was checked only in specific cases and there
was no universal check for that before applying a block. This could have
resulted into a situation, when a block with a rewards distribution that
cannot be calculated could have been applied. Thus, a node might have
been stuck in an invalid state.,

A particular example of that is a block signed by the same node that is
being retired by a transaction in this block. Because a node is already
retired when this block is applied, the rewards for this block cannot be
distributed.